Module type Board.GameBoard

type t

type t' = {

board_type : t;

added : bool;

}

type t'' = {

board_type : t;

shot : bool;

ship_shot : Ship.AShip.t option;

}

type occupy =

| Ocean

| Ship of Ship.AShip.t

| Hit

| Miss

val set_up_board : occupy -> t

val check_valid : t -> Ship.AShip.t -> int -> int -> int -> int -> bool

val place_ship : t -> Ship.AShip.t -> int -> int -> int -> int -> t'

val shoot : t -> int -> int -> t''

Given x and y coordinates this function searches to see if there is a ship at that position in the graph and if so shoots/hits that part of the ship shoot x y Requires: That the position x,y is a valid position in the graph Returns: True if part of a ship is at the inputted positon (x,y) in the GameBoard and False otherwise

val get_board : t -> occupy array array

val get_pos : t -> int -> int -> occupy