Module type Computer.comp

type t

val create_ai : t

val set_board_ai : t -> unit

val get_board : t -> Board.BattleGround.t

val set_board : t -> Board.BattleGround.t -> unit

val get_bag : t -> ShipsBag.PlayerList.t

val rand_Move : t -> Board.BattleGround.t -> Board.BattleGround.t''

rand_Move playerBoard Returns: a new board with randomized shoot locations. If the shot hits, the ship and board is updated. Otherwise only the board is updated.

val rand1_Move : t -> Board.BattleGround.t -> Board.BattleGround.t''

rand_Move playerBoard Returns: a new board with a shot fired based on current hit and miss locations. If the shot hits, the ship and board is updated. Otherwise only the board is updated.